Moved button in UiKitSample in debug build
