+++ app/src/main/java/care/intouch/app/UiKitSample.kt
@@ -1,4 +1,4 @@
-package care.intouch.app.ui.theme
+package care.intouch.app
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
